- Changed grade type

+++ src/main/kotlin/com/amalip/data/entity/UserCourseEntity.kt
@@ -1,6 +1,7 @@
 package com.amalip.data.entity
 
 import org.ktorm.schema.Table
+import org.ktorm.schema.double
 import org.ktorm.schema.int
 
 object UserCourseEntity : Table<Nothing>("user_course") {
@@ -8,8 +9,8 @@ object UserCourseEntity : Table<Nothing>("user_course") {
     val id = int("id").primaryKey()
     val courseId = int("id_course")
     val userId = int("id_user")
-    val grade1 = int("grade1")
-    val grade2 = int("grade2")
-    val grade3 = int("grade3")
+    val grade1 = double("grade1")
+    val grade2 = double("grade2")
+    val grade3 = double("grade3")
 
 }
